fix(arc-624): Do not require login for content pages (404)
