+++ src/modules/home/components/RequestAccessBlade/RequestAccessBlade.tsx
@@ -6,7 +6,7 @@ import { Controller, useForm } from 'react-hook-form';
 import { StringParam, useQueryParams } from 'use-query-params';
 
 import { VISITOR_SPACE_SLUG_QUERY_KEY } from '@home/const';
-import { useGetReadingRoom } from '@reading-room/hooks/get-reading-room';
+import { useGetVisitorSpace } from '@reading-room/hooks/get-reading-room';
 import { Blade, Icon, SpacePreview } from '@shared/components';
 import { OPTIONAL_LABEL } from '@shared/const';
 
@@ -19,7 +19,7 @@ const RequestAccessBlade: FC<RequestAccessBladeProps> = ({ onSubmit, ...bladePro
 	const [query] = useQueryParams({
 		[VISITOR_SPACE_SLUG_QUERY_KEY]: StringParam,
 	});
-	const { data: space } = useGetReadingRoom(query[VISITOR_SPACE_SLUG_QUERY_KEY]);
+	const { data: space } = useGetVisitorSpace(query[VISITOR_SPACE_SLUG_QUERY_KEY]);
 
 	const {
 		control,
